RequirementMachine: Move diagnostics code into a new Diagnostics.cpp
