+++ lib/AST/RequirementMachine/RequirementLowering.h
@@ -56,25 +56,6 @@ void realizeInheritedRequirements(TypeDecl *decl, Type type,
                                   SmallVectorImpl<StructuralRequirement> &result,
                                   SmallVectorImpl<RequirementError> &errors);
 
-/// Policy for the fixit that transforms 'T : S' where 'S' is not a protocol
-/// or a class into 'T == S'.
-enum AllowConcreteTypePolicy {
-  /// Any type parameter can be concrete.
-  All,
-
-  /// Only associated types can be concrete.
-  AssocTypes,
-
-  /// Only nested associated types can be concrete. This is for protocols,
-  /// where we don't want to suggest making an associated type member of
-  /// 'Self' concrete.
-  NestedAssocTypes
-};
-
-bool diagnoseRequirementErrors(ASTContext &ctx,
-                               ArrayRef<RequirementError> errors,
-                               AllowConcreteTypePolicy concreteTypePolicy);
-
 // Defined in ConcreteContraction.cpp.
 bool performConcreteContraction(
     ArrayRef<StructuralRequirement> requirements,
